fix(infra): S163 — 5 root fixes: make_interval cast, EXIT side, uPnL persist, integrity check, idle-in-tx
1. resolution_backfill.py: Add ::int cast to power() result in make_interval().
   PG16 rejects float for hours param. Exponential backoff + dead-letter
   promotion have been inert since deployment.

2. paper_trading.py: EXIT events now record token-outcome side (YES/NO)
   instead of hardcoded "SELL". Fixes data model inconsistency where
   entries and exits for the same position had different side values.

3. position_manager.py: Explicit SQL UPDATE for current_price/unrealized_pnl
   on expunged positions. S161 expunge() detached ORM objects from session,
   making attribute changes invisible to session.commit(). 52 positions
   (44 Weather + 8 Esports) had frozen uPnL=$0, degrading exit decisions.

4. bot_pnl.py: Integrity check groups by market_id only (not side).
   Historical SELL-side EXIT events caused 3,083+ false warnings.

5. prune_market_prices.py: Hourly cleanup of idle-in-transaction (aborted)
   PG connections older than 120s. Self-healing for session poisoning residue.

1789 tests pass. All 15 bots affected by Fix 2-3.

+++ base_engine/data/resolution_backfill.py
@@ -336,7 +336,7 @@ async def run_resolution_backfill(
                     await _lc.execute(text("""
                         UPDATE traded_markets
                         SET last_checked_at = NOW() + make_interval(
-                            hours => LEAST(power(3, LEAST(COALESCE(check_fail_count, 0), 4)), 168)
+                            hours => LEAST(power(3, LEAST(COALESCE(check_fail_count, 0), 4)), 168)::int
                         ),
                         check_fail_count = COALESCE(check_fail_count, 0) + 1,
                         resolution_status = CASE
